src/requirements.txt
langchain-groq 
langgraph 
tavily-python 
google-api-python-client 
langchain-community  
beautifulsoup4 
     